chore: upload image with Cloudinary and fix bugs

+++ server/src/modules/equipments/equipment-service.ts
@@ -1,4 +1,7 @@
 import createHttpError from 'http-errors';
+import { Types } from 'mongoose';
+
+import { uploadImage } from '~/utils/cloudinary';
 
 import EquipmentModel from './equipment-model';
 import { IEquipment } from './equipment-type';
@@ -12,15 +15,26 @@ const EquipmentService = {
 
     return equipments;
   },
+
   findById: async (equipmentId: string) => {
+    if (!Types.ObjectId.isValid(equipmentId)) {
+      throw createHttpError(400, 'Invalid ObjectId');
+    }
+
     const equipment = await EquipmentModel.findById(equipmentId);
+
     if (!equipment) {
       throw createHttpError(404, 'Equipment not found');
     }
 
     return equipment;
   },
-  create: async (equipmentData: IEquipment) => {
+
+  create: async (equipmentData: IEquipment, file?: Express.Multer.File) => {
+    if (!file) {
+      return createHttpError(400, 'No file provided');
+    }
+
     const existingEquipment = await EquipmentModel.findOne({
       title: equipmentData.title
     });
@@ -28,22 +42,63 @@ const EquipmentService = {
       throw createHttpError(409, 'Equipment with this title already exists');
     }
 
-    const newEquipment = EquipmentModel.create(equipmentData);
+    const uploadResult = await uploadImage(file.buffer);
+
+    if (!uploadResult.success || !uploadResult.data) {
+      throw createHttpError(
+        500,
+        uploadResult.error || 'Failed to upload image'
+      );
+    }
+
+    const imageUrl = uploadResult.data.secure_url;
+
+    const newEquipmentData = { ...equipmentData, image: imageUrl };
+
+    const newEquipment = await EquipmentModel.create(newEquipmentData);
     if (!newEquipment) {
       throw createHttpError(500, 'Failed to create equipment');
     }
 
     return newEquipment;
   },
-  update: async (equipmentId: string, equipmentData: Partial<IEquipment>) => {
-    const equipment = await EquipmentModel.findById(equipmentId);
-    if (!equipment) {
+
+  update: async (
+    equipmentId: string,
+    equipmentData: Partial<IEquipment>,
+    file?: Express.Multer.File
+  ) => {
+    if (!Types.ObjectId.isValid(equipmentId)) {
+      throw createHttpError(400, 'Invalid ObjectId');
+    }
+
+    const existingEquipment = await EquipmentModel.findById(equipmentId);
+    if (!existingEquipment) {
       throw createHttpError(404, 'Equipment not found');
     }
 
+    let imageUrl: string | undefined;
+    if (file) {
+      const uploadResult = await uploadImage(file.buffer);
+
+      if (!uploadResult.success || !uploadResult.data) {
+        throw createHttpError(
+          500,
+          uploadResult.error || 'Failed to upload image'
+        );
+      }
+
+      imageUrl = uploadResult.data.secure_url;
+    }
+
+    const updatedEquipmentData = {
+      ...equipmentData,
+      image: imageUrl || existingEquipment.image
+    };
+
     const updatedEquipment = await EquipmentModel.findByIdAndUpdate(
       equipmentId,
-      equipmentData,
+      updatedEquipmentData,
       {
         new: true,
         runValidators: true
@@ -56,8 +111,14 @@ const EquipmentService = {
 
     return updatedEquipment;
   },
+
   remove: async (equipmentId: string) => {
+    if (!Types.ObjectId.isValid(equipmentId)) {
+      throw createHttpError(400, 'Invalid ObjectId');
+    }
+
     const equipment = await EquipmentModel.findByIdAndDelete(equipmentId);
+
     if (!equipment) {
       throw createHttpError(404, 'Equipment not found');
     }
